Use common file for collector testing This main functino will serve all collectors
