+++ collectors/collector_test.go
@@ -1,0 +1,43 @@
+package collectors
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/wanguard-api/v1/license_manager", func(w http.ResponseWriter, r *http.Request) {
+		w.Write([]byte(licenseManagerPayload()))
+	})
+
+	server := httptest.NewServer(mux)
+	defer server.Close()
+
+	os.Setenv("TEST_SERVER_URL", server.URL)
+	exitCode := m.Run()
+	os.Unsetenv("TEST_SERVER_URL")
+
+	os.Exit(exitCode)
+}
+
+func licenseManagerPayload() string {
+	return `{
+  "software_version": "8.3-21",
+  "licensed_sensors": 1,
+  "licensed_sensors_used": 1,
+  "licensed_sensors_remaining": 0,
+  "licensed_sensor_interfaces": "∞",
+  "licensed_dpdk_engines": 0,
+  "licensed_dpdk_engines_used": 0,
+  "licensed_dpdk_engines_remaining": 0,
+  "licensed_filters": 1,
+  "licensed_filters_used": 1,
+  "licensed_filters_remaining": 0,
+  "licensed_on": "2024-10-24 12:15:29",
+  "license_expiry_date_remaining": "365 days",
+  "support_expiry_date_remaining": "365 days"
+}`
+}
